requirements.txt
anyio==3.6.2
asyncio==3.4.3
bidict==0.22.1
click==8.1.3
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
netifaces==0.10.6
pydantic==1.10.7
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
